(PC-18519)[PRO] refactor: extract duplicate function to external file
