+++ pro/src/core/OfferEducational/utils/index.ts
@@ -4,6 +4,7 @@ export * from './hasStatusCode'
 export * from './filterEducationalCategories'
 export * from './setInitialFormValues'
 export * from './createStockDataPayload'
+export * from './createOfferFromTemplate'
 export * from './createPatchStockDataPayload'
 export * from './buildDatetimesForStockPayload'
 export * from './extractOfferIdAndOfferTypeFromRouteParams'
