+++ pro/src/pages/Offers/Offers/OfferItem/Cells/DuplicateOfferCell/DuplicateOfferCell.tsx
@@ -5,9 +5,10 @@ import {
   Events,
   OFFER_FROM_TEMPLATE_ENTRIES,
 } from 'core/FirebaseEvents/constants'
-import getCollectiveOfferFormDataApdater from 'core/OfferEducational/adapters/getCollectiveOfferFormDataAdapter'
-import getCollectiveOfferTemplateAdapter from 'core/OfferEducational/adapters/getCollectiveOfferTemplateAdapter'
-import postCollectiveOfferAdapter from 'core/OfferEducational/adapters/postCollectiveOfferAdapter'
+import {
+  createOfferFromTemplate,
+  oldCreateOfferFromTemplate,
+} from 'core/OfferEducational'
 import useActiveFeature from 'hooks/useActiveFeature'
 import useAnalytics from 'hooks/useAnalytics'
 import useNotification from 'hooks/useNotification'
@@ -20,50 +21,6 @@ import styles from '../../OfferItem.module.scss'
 import DuplicateOfferDialog from './DuplicateOfferDialog'
 
 export const LOCAL_STORAGE_HAS_SEEN_MODAL_KEY = 'DUPLICATE_OFFER_MODAL_SEEN'
-
-// TODO (atrancart): delete when feature flag is removed
-export const oldCreateOfferFromTemplate = async (
-  history: ReturnType<typeof useHistory>,
-  templateOfferId: string
-) => {
-  history.push(`/offre/duplication/collectif/${templateOfferId}`)
-}
-
-export const createOfferFromTemplate = async (
-  history: ReturnType<typeof useHistory>,
-  notify: ReturnType<typeof useNotification>,
-  templateOfferId: string
-) => {
-  const offerTemplateResponse = await getCollectiveOfferTemplateAdapter(
-    templateOfferId
-  )
-
-  if (!offerTemplateResponse.isOk) {
-    return notify.error(offerTemplateResponse.message)
-  }
-
-  const result = await getCollectiveOfferFormDataApdater({
-    offererId: offerTemplateResponse.payload.venue.managingOffererId,
-    offer: offerTemplateResponse.payload,
-  })
-
-  if (!result.isOk) {
-    notify.error(result.message)
-  }
-
-  const { initialValues } = result.payload
-
-  const { isOk, message, payload } = await postCollectiveOfferAdapter({
-    offer: initialValues,
-    offerTemplateId: templateOfferId,
-  })
-
-  if (!isOk) {
-    return notify.error(message)
-  }
-
-  history.push(`/offre/collectif/${payload.id}/creation`)
-}
 
 const DuplicateOfferCell = ({
   isTemplate,
